+++ RELEASE_NOTES.md
@@ -1,3 +1,10 @@
+### Version 1.0.0
+__Changes__
+- handle gene annotation functions from Genomes >= 9.0 (Features >= 3.0)
+- support FeatureSets with AnnotatedMetagenomeAssembly features
+- added Github Actions testing
+- made all test data uploaded
+
 ### Version 0.0.9
 __Changes__
 - Added genome_ref_to_obj_name return value